Move kfp ack sm test to test_sanity
