+++ tests/e2e/tests/test_sanity_portforward.py
@@ -91,47 +91,6 @@ def wait_for_run_succeeded(kfp_client, run, job_name, pipeline_id):
 
     return wait_for(callback, timeout=900)
 
-
-@pytest.fixture(scope="class")
-def sagemaker_execution_role(region, metadata, request):
-    sagemaker_execution_role_name = "role-" + RANDOM_PREFIX
-    managed_policies = ["arn:aws:iam::aws:policy/AmazonS3FullAccess", "arn:aws:iam::aws:policy/AmazonSageMakerFullAccess"]
-    role = IAMRole(name=sagemaker_execution_role_name, region=region, policy_arns=managed_policies)
-    metadata_key = "sagemaker_execution_role"
-
-    resource_details = {}
-
-    def on_create():
-        trust_policy = {
-            "Version": "2012-10-17",
-            "Statement": [
-                {
-                    "Effect": "Allow",
-                    "Principal": {"Service": "sagemaker.amazonaws.com"},
-                    "Action": "sts:AssumeRole",
-                }
-            ],
-        }
-
-        sagemaker_execution_role_arn = role.create(
-            policy_document=json.dumps(trust_policy)
-        )
-
-        resource_details["name"] = sagemaker_execution_role_name
-        resource_details["arn"] = sagemaker_execution_role_arn
-
-    def on_delete():
-        role.delete()
-
-    return configure_resource_fixture(
-        metadata=metadata,
-        request=request,
-        resource_details=resource_details,
-        metadata_key=metadata_key,
-        on_create=on_create,
-        on_delete=on_delete,
-    )
-
 @pytest.fixture(scope="class")
 def s3_bucket_with_data(region):
     bucket_name = "s3-" + RANDOM_PREFIX
@@ -142,12 +101,6 @@ def s3_bucket_with_data(region):
     yield
     bucket.delete()
 
-@pytest.fixture(scope="class")
-def clean_up_training_jobs_in_user_ns():
-    yield 
-
-    cmd = f"kubectl delete trainingjobs --all -n {DEFAULT_USER_NAMESPACE}".split()
-    subprocess.Popen(cmd)
 
 class TestSanity:
     @pytest.fixture(scope="class")
@@ -279,44 +232,3 @@ class TestSanity:
         # The second condition is now required in case the kfp test runs before this one.
         assert expected_output in output or "training-job-" in output
     
-    def test_run_kfp_sagemaker_pipeline(
-        self, region, metadata, s3_bucket_with_data, sagemaker_execution_role, kfp_client, clean_up_training_jobs_in_user_ns
-    ):
-
-        experiment_name = "experiment-" + RANDOM_PREFIX
-        experiment_description = "description-" + RANDOM_PREFIX
-        sagemaker_execution_role_name = "role-" + RANDOM_PREFIX
-        bucket_name = "s3-" + RANDOM_PREFIX
-        job_name = "kfp-run-" + RANDOM_PREFIX
-
-        sagemaker_execution_role_details = metadata.get("sagemaker_execution_role")
-        sagemaker_execution_role_arn = sagemaker_execution_role_details["arn"]
-
-        
-        experiment = kfp_client.create_experiment(
-            experiment_name,
-            description=experiment_description,
-            namespace=DEFAULT_USER_NAMESPACE,
-        )
-
-        pipeline_id = kfp_client.get_pipeline_id(PIPELINE_NAME_KFP)
-
-        params = {
-            "sagemaker_role_arn": sagemaker_execution_role_arn,
-            "s3_bucket_name": bucket_name,
-        }
-
-        run = kfp_client.run_pipeline(
-            experiment.id, job_name=job_name, pipeline_id=pipeline_id, params=params
-        )
-
-        assert run.name == job_name
-        assert run.pipeline_spec.pipeline_id == pipeline_id
-        assert run.status == None
-
-        wait_for_run_succeeded(kfp_client, run, job_name, pipeline_id)
-
-        kfp_client.delete_experiment(experiment.id)
-        
-        cmd = "kubectl delete trainingjobs --all -n kubeflow-user-example-com".split()
-        subprocess.Popen(cmd)
